requirements.txt
smsir-python
Django
MySQL
requests